application de gestion de vote

+++ app/Http/Controllers/RegionController.php
@@ -3,11 +3,28 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use App\Models\Region;
 use App\Models\Republique;
 
 class RegionController extends Controller
 {
+
+  public function __construct()
+  {
+    try
+    {
+      $this->middleware('auth');//->except(['index'])
+      $this->middleware('limite');
+    }
+    catch(\Exception $exception)
+    {
+        // return $exception->getMessage();
+        return redirect('erreur')->with('messagealert',$exception->getMessage());
+    }
+  }
+
+  
             /**
      * Display a listing of the resource.
      *
@@ -16,10 +33,18 @@ class RegionController extends Controller
    // Afficher les republiques
    public function index()
    {
-   $regions = Region::all();
-   $republiques = Republique::all();
+     try
+     {
+        $regions = Region::all();
+        $republiques = Republique::all();
 
-   return view('Region.index', compact('regions', 'republiques'));
+        return view('Region.index', compact('regions', 'republiques'));
+     }
+    catch(\Exception $exception)
+     {
+        // return $exception->getMessage();
+        return redirect('erreur')->with('messagealert',$exception->getMessage());
+     }
    }
 
      /**
@@ -30,10 +55,18 @@ class RegionController extends Controller
 
    public function create()
    {
+     try
+     {
        $region = new Region();
        $republiques = Republique::all();
 
        return view('Region.create',compact('region', 'republiques'));
+     }
+    catch(\Exception $exception)
+    {
+          // return $exception->getMessage();
+          return redirect('erreur')->with('messagealert',$exception->getMessage());
+    }
    }
 
      /**
@@ -45,9 +78,17 @@ class RegionController extends Controller
 
    public function store()
    {   
+     try
+     {
        $regions = Region::create($this->validator());
 
        return redirect('regions')->with('message', 'Région bien ajoutée.');
+     }
+      catch(\Exception $exception)
+      {
+            // return $exception->getMessage();
+            return redirect('erreur')->with('messagealert',$exception->getMessage());
+      }
    }
 
     /**
@@ -59,7 +100,15 @@ class RegionController extends Controller
 
    public function show(Region $region)
    {
-     return view('Region.show',compact('region'));
+     try
+     {
+       return view('Region.show',compact('region'));
+     }
+     catch(\Exception $exception)
+     {
+          // return $exception->getMessage();
+          return redirect('erreur')->with('messagealert',$exception->getMessage());
+     }
    }
 
   /**
@@ -71,8 +120,16 @@ class RegionController extends Controller
 
    public function edit(Region $region)
    {
+     try
+     {
        $republiques = Republique::all();
        return view('Region.edit', compact('region', 'republiques'));
+     }
+      catch(\Exception $exception)
+      {
+            // return $exception->getMessage();
+            return redirect('erreur')->with('messagealert',$exception->getMessage());
+      }
    }
 
       /**
@@ -85,9 +142,17 @@ class RegionController extends Controller
 
    public function update(Region $region)
    {
+     try
+     {
       $region->update($this->validator());
 
        return redirect('regions/' . $region->id);
+     }
+      catch(\Exception $exception)
+      {
+            // return $exception->getMessage();
+            return redirect('erreur')->with('messagealert',$exception->getMessage());
+      }
    }
 
     /**
@@ -99,9 +164,27 @@ class RegionController extends Controller
 
    public function destroy(Region $region)
    {
-    $region->delete();
+     try
+     {
+        $reference = DB::table('users')
+        ->where('users.region_id','=',$region->id)
+        ->select('users.id')
+        ->first();
 
-       return redirect('regions');
+        if($reference->id == null)
+        {
+          $region->delete();
+
+          return redirect('regions');
+        }
+         return redirect('regions')->with('messagealert','Cette région est referencée dans une autre table');
+
+     }
+      catch(\Exception $exception)
+      {
+            // return $exception->getMessage();
+            return redirect('erreur')->with('messagealert',$exception->getMessage());
+      }
    }
 
    private  function validator()
